Added list of characters for avengers and xmen
Started two lists.  Avengers file does not use all the marvel movies to
draw characters from.  it only uses the actual avengers movie as the
defining points for when characters are on the list.

The xmen list goes by all the xmen movie but does the same thing. again,
it's only for main characters.